Brain Tumor Project/requirements.txt.txt
gradio
pillow
torch